Fixed datatypes

+++ SolarController/solarcontroller/RenogyData.h
@@ -18,20 +18,20 @@ public:
     uint8_t panelPower;
     float loadVoltage;                     // volts
     float loadCurrent;                     // amps
-    uint8_t loadPower;                     // watts
+    uint16_t loadPower;                     // watts
     float minBatteryVoltageToday;          // volts
     float maxBatteryVoltageToday;          // volts
     float maxChargingCurrentToday;         // amps
     float maxDischargingCurrentToday;      // amps
-    uint8_t maxChargePowerToday;           // watts
-    uint8_t maxDischargePowerToday;        // watts
-    uint8_t chargeAmphoursToday;           // amp hours
-    uint8_t dischargeAmphoursToday;        // amp hours
-    uint8_t powerGenerationToday;          // watt hours
-    uint8_t powerConsumptionToday;         // watt hours
-    uint8_t totalNumOperatingDays;         // days
-    uint8_t totalNumBatteryOverDischarges; // count
-    uint8_t totalNumBatteryFullCharges;    // count
+    uint16_t maxChargePowerToday;           // watts
+    uint16_t maxDischargePowerToday;        // watts
+    uint16_t chargeAmphoursToday;           // amp hours
+    uint16_t dischargeAmphoursToday;        // amp hours
+    uint16_t powerGenerationToday;          // watt hours
+    uint16_t powerConsumptionToday;         // watt hours
+    uint16_t totalNumOperatingDays;         // days
+    uint16_t totalNumBatteryOverDischarges; // count
+    uint16_t totalNumBatteryFullCharges;    // count
     uint32_t totalChargingAmphours;        // amp hours
     uint32_t totalDischargingAmphours;     // amp hours
     uint32_t cumulativePowerGeneration;    // kW hours
